Security fix for python interpreter replacement
