+++ lib/hardware_info.py
@@ -6,69 +6,13 @@ import re
 import os
 import platform
 import pprint
-import subprocess
 import psutil
 import sys
+from lib.hardware_info_root_original import rdr, rpwr, rfwr, cf, get_values
 
 REGEX_PARAMS = re.MULTILINE | re.IGNORECASE
 
 CURRENT_PATH = os.path.dirname(__file__)
-
-def call_function(func, arguments=None, attribute=None):
-    if arguments is None:
-        arguments = []
-    if attribute:
-        return getattr(func(*arguments), attribute)
-    return func(*arguments)
-
-def read_file_with_regex(file, regex, params=REGEX_PARAMS):
-    '''Reads the content of a file and then tries to match the regex and returns the match described by 'o' '''
-
-    try:
-        with open(file, 'r', encoding='utf8') as file_ptr:
-            file_data = file_ptr.read()
-    except FileNotFoundError:
-        return 'File not found'
-
-    match = re.search(regex, file_data, params)
-
-    return match.group('o') if match is not None else 'Unknown'
-
-
-def read_process_with_regex(path, regex, params=REGEX_PARAMS):
-    '''Reads the data from a process and then matches the output. The process must terminate and not require user
-       input! The matching character for the regex is a 'o'. If the process fails (exit val not 0) an exception
-       is thrown.'''
-    result = subprocess.run(path, stdout=subprocess.PIPE, stderr=subprocess.PIPE,
-        shell=True, encoding='UTF-8', check=False)
-    if result.returncode != 0:
-        return 'Unknown'
-
-    match = re.search(regex, result.stdout, params)
-
-    return match.group('o') if match is not None else 'Unknown'
-
-
-def read_directory_recursive(directory):
-    '''Reads all the files in a directory recursively and adds them to the return dict. We ignore most errors.'''
-    output_dic = {}
-
-    dir_path = directory
-    for (dir_path, _, file_names) in os.walk(dir_path):
-        for filename in file_names:
-            try:
-                with open(os.path.join(dir_path, filename), 'r', encoding='utf8') as file_ptr:
-                    output_dic[os.path.join(dir_path, filename)] = file_ptr.read()
-            except (FileNotFoundError, PermissionError, OSError):
-                continue
-    return output_dic
-
-
-# Defining shortcuts to make the lines shorter
-rfwr = read_file_with_regex
-rpwr = read_process_with_regex
-rdr = read_directory_recursive
-cf = call_function
 
 # pylint: disable=unnecessary-lambda-assignment
 # read_process_with_regex with duplicates removed (single)
@@ -103,7 +47,7 @@ linux_info_list = [
         re.IGNORECASE | re.DOTALL,
     ],
     [rfwr, 'Hyper Threading', '/sys/devices/system/cpu/smt/active', r'(?P<o>.*)'],
-    [rdr, 'CPU complete dump', '/sys/devices/system/cpu/'],
+    [rdr, 'CPU Complete Dump', '/sys/devices/system/cpu/'],
     # This is also listed in the complete dump but we include it here again so it is more visible in the listing
 
     # Note that this information is only accurate if intel_pstate is in active mode
@@ -115,7 +59,7 @@ linux_info_list = [
     [rfwr, 'Turbo Boost (Legacy non intel_pstate)', '/sys/devices/system/cpu/cpufreq/boost', r'(?P<o>.*)'],
     [rfwr, 'Virtualization', '/proc/cpuinfo', r'(?P<o>hypervisor)'],
     [rpwrs, 'SGX', f"{os.path.join(CURRENT_PATH, '../tools/sgx_enable')} -s", r'(?P<o>.*)', re.IGNORECASE | re.DOTALL],
-    [rfwr, 'IO scheduling', '/sys/block/sda/queue/scheduler', r'(?P<o>.*)'],
+    [rfwr, 'IO Scheduling', '/sys/block/sda/queue/scheduler', r'(?P<o>.*)'],
     [rpwr, 'Network Interfaces', 'ip addr | grep ether -B 1', r'(?P<o>.*)', re.IGNORECASE | re.DOTALL],
     [rfwr, 'Current Clocksource', '/sys/devices/system/clocksource/clocksource0/current_clocksource', r'(?P<o>.*)'],
 
@@ -147,11 +91,6 @@ def get_list():
         return mac_info_list
 
     return linux_info_list
-
-
-def get_values(list_of_tasks):
-    '''Creates an object with all the data populated'''
-    return {x[1]: x[0](*x[2:]) for x in list_of_tasks}
 
 def get_default_values():
     return get_values(get_list())
